Validación
Contiene validacion de campo en financiamiento y funcional todos los
ingresos excepto impuestos, no se como hacer

+++ app/database/migrations/2015_07_12_213520_crear_tabla_cantidades.php
@@ -17,9 +17,9 @@ class CrearTablaCantidades extends Migration {
 			$table->increments('id');
 			$table->date('mesCant');
 			$table->integer('unidadMes');
-			$table->decimal('precioMes', 10, 2);
+			$table->double('precioMes', 10, 2);
 			$table->string('producto',30);
-			$table->decimal('total', 10, 2);
+			$table->double('total', 10, 2);
 			
 			$table-> unsignedInteger('id_ingresos');
 			$table-> foreign('id_ingresos')->references('id')->on('tbl_ingresos');
